requirements.txt
asgiref==3.2.7
autopep8==1.6.0
Django==3.2.10
django-model-utils==4.0.0
html5lib==1.0.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.8.0
PyPDF2==1.26.0
python-dotenv==0.19.2
pytz==2020.1
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
Unipath==1.1
webencodings==0.5.1
xhtml2pdf==0.2.4
